# api/openapi.yaml
openapi: 3.0.3
info:
  title: WhereToLive HTTP API
  version: 0.1.0
  description: Authoritative HTTP contract for the WhereToLive modular monolith.
servers:
  - url: /api
paths:
...
  /public/auth/register:
    post:
      operationId: registerConsumer
      tags: [public-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ConsumerRegistrationRequest"}
      responses:
        "202": {description: Request accepted regardless of account existence or delivery enqueue outcome}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/resend-verification:
    post:
      operationId: resendConsumerVerification
      tags: [public-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ConsumerEmailRequest"}
      responses:
        "202": {description: Request accepted regardless of account existence or delivery enqueue outcome}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/recover:
    post:
      operationId: requestConsumerRecovery
      tags: [public-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ConsumerEmailRequest"}
      responses:
        "202": {description: Request accepted regardless of account existence or delivery enqueue outcome}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/verify-email:
    post:
      operationId: verifyConsumerEmail
      tags: [public-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ConsumerCodeRequest"}
      responses:
        "204": {description: Completed}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/reset-password:
    post:
      operationId: resetConsumerPassword
      tags: [public-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ConsumerResetRequest"}
      responses:
        "204": {description: Completed}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/login:
    post:
      operationId: loginConsumer
      tags: [public-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ConsumerLoginRequest"}
      responses:
        "200":
          description: Consumer access token and isolated refresh cookies
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AccessTokenResponse"}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/refresh:
    post:
      operationId: refreshConsumer
      tags: [public-auth]
      security: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Consumer access token and isolated refresh cookies
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AccessTokenResponse"}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/logout:
    post:
      operationId: logoutConsumer
      tags: [public-auth]
      security: [{consumerBearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": {description: Completed}
        "400": {$ref: "#/components/responses/PublicInvalidQuery"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "403": {description: Challenge or CSRF rejected}
        "429": {description: Request budget exhausted}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
  /public/auth/me:
    get:
      operationId: getConsumerProfile
      tags: [public-auth]
      security: [{consumerBearerAuth: []}]
      responses:
        "200":
          description: Current consumer profile
          content:
            application/json:
              schema: {$ref: "#/components/schemas/UserProfile"}
        "401": {$ref: "#/components/responses/AuthenticationFailed"}
        "500": {$ref: "#/components/responses/PublicInternalError"}
        "503": {$ref: "#/components/responses/PublicUnavailable"}
...
  /auth/refresh:
    post:
      operationId: refresh
      summary: Rotate the refresh secret and issue an access token
      tags: [authentication]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Session rotated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
...
components:
  securitySchemes:
    consumerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    platformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
...
  responses:
    PublicInvalidQuery:
      description: Invalid query
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PublicUnavailable:
      description: Catalogue is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PublicInternalError:
      description: Internal error without private details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    AuthenticationFailed:
      description: Tenant or credentials are invalid, or the session is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    ConsumerRegistrationRequest:
      type: object
      additionalProperties: false
      required: [email, password, challenge]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, minLength: 12, maxLength: 1024}
        challenge: {type: string, minLength: 1, maxLength: 2048}
    ConsumerLoginRequest:
      type: object
      additionalProperties: false
      required: [email, password, challenge]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, minLength: 1, maxLength: 1024}
        challenge: {type: string, minLength: 1, maxLength: 2048}
    ConsumerEmailRequest:
      type: object
      additionalProperties: false
      required: [email, challenge]
      properties:
        email: {type: string, format: email, maxLength: 254}
        challenge: {type: string, minLength: 1, maxLength: 2048}
    ConsumerCodeRequest:
      type: object
      additionalProperties: false
      required: [code, challenge]
      properties:
        code: {type: string, minLength: 32, maxLength: 128}
        challenge: {type: string, minLength: 1, maxLength: 2048}
    ConsumerResetRequest:
      type: object
      additionalProperties: false
      required: [code, password, challenge]
      properties:
        code: {type: string, minLength: 32, maxLength: 128}
        password: {type: string, minLength: 12, maxLength: 1024}
        challenge: {type: string, minLength: 1, maxLength: 2048}
...
    UserProfile:
      type: object
      additionalProperties: false
      required: [id, username, status, updatedAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string, format: email, nullable: true }
        status: { type: string, enum: [active] }
        updatedAt: { type: string, format: date-time }
...
    AccessTokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresIn, csrfToken]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        csrfToken:
          type: string
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        traceId:
          type: string